requirements.txt
asgiref==3.7.2
cymysql==0.9.18
Django==4.2.2
django-cymysql==3.2.0
django-mysql==4.9.0
django-resized==1.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
xlrd==2.0.1
xlrd2==1.3.4
